sync 843c6ce44ce4f

+++ include/ahci.h
@@ -137,8 +137,8 @@ struct ahci_ioports {
 	void __iomem	*port_mmio;
 	struct ahci_cmd_hdr	*cmd_slot;
 	struct ahci_sg		*cmd_tbl_sg;
-	ulong	cmd_tbl;
-	u32	rx_fis;
+	uintptr_t	cmd_tbl;
+	uintptr_t	rx_fis;
 };
 
 /**
